requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pkg-resources==0.0.0
python-dotenv==0.17.1
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
